Move some update from let_internal_branch to main

+++ rclc_exp/src/exp_single_thread_single_core.c
@@ -335,7 +335,7 @@ int main(int argc, char const *argv[])
     {
       RCCHECK(rclc_executor_set_semantics(&executor1, RCLCPP_EXECUTOR));
     }
-    printf("ExecutorID Executor1 %d\n", &executor1);
+    printf("ExecutorID Executor1 %lu\n", (unsigned long) &executor1);
 
     ////////////////////////////////////////////////////////////////////////////
     // Configuration of Linux threads
